+++ NEWS.md
@@ -151,6 +151,15 @@ UPDATE events
 UPDATE events
    SET "classification.taxonomy" = 'abusive content', "classification.type" = 'spam', "classification.identifier" = 'openrelay'
    WHERE "malware.name" = 'openrelay' AND "feed.name" = "Spamhaus CERT";
+UPDATE events
+   SET "protocol.application" = 'portmapper'
+   WHERE "classification.identifier" = 'openportmapper' AND "feed.name" = "Open-Portmapper";
+UPDATE events
+   SET "protocol.application" = 'netbios-nameservice'
+   WHERE "classification.identifier" = 'opennetbios' AND "feed.name" = "Open-NetBIOS-Nameservice";
+UPDATE events
+   SET "protocol.application" = 'ipsec'
+   WHERE "classification.identifier" = 'openike' AND "feed.name" = "Vulnerable-ISAKMP";
 ```
 
 1.0.3 Bugfix release (2018-02-05)