+++ NEWS.md
@@ -3,7 +3,7 @@ NEWS
 
 See the changelog for a full list of changes.
 
-1.0.4 Bugfix release (unreleased)
+1.0.4 Bugfix release (2018-04-20)
 ---------------------------------
 
 ### Postgres databases